Divide the test helpers into two files
