+++ lib/warden/test/helpers.rb
@@ -1,7 +1,5 @@
 # encoding: utf-8
 
-require 'rack'
-
 module Warden
   module Test
     # A collection of test helpers for testing full stack rack applications using Warden
@@ -33,59 +31,6 @@ module Warden
           proxy.logout(*scopes)
         end
       end
-
-      # A helper method that provides the warden object by mocking the env variable.
-      # @api public
-      def warden
-        @warden ||= begin
-          env['warden']
-        end
-      end
-
-      private
-
-      def env
-        @env ||= begin
-          request = Rack::MockRequest.env_for(
-            "/?#{Rack::Utils.build_query({})}",
-            { 'HTTP_VERSION' => '1.1', 'REQUEST_METHOD' => 'GET' }
-          )
-          app.call(request)
-
-          request
-        end
-      end
-
-      def app
-        @app ||= begin
-          opts = {
-            failure_app: lambda {
-              [401, { 'Content-Type' => 'text/plain' }, ['You Fail!']]
-            },
-            default_strategies: :password,
-            default_serializers: :session
-          }
-          Rack::Builder.new do
-            use Warden::Test::Helpers::Session
-            use Warden::Manager, opts, &proc {}
-            run lambda { |e|
-              [200, { 'Content-Type' => 'text/plain' }, ['You Win']]
-            }
-          end
-        end
-      end
-
-      class Session
-        attr_accessor :app
-        def initialize(app,configs = {})
-          @app = app
-        end
-
-        def call(e)
-          e['rack.session'] ||= {}
-          @app.call(e)
-        end
-      end # session
     end
   end
 end
