+++ spec/spec_helper.rb
@@ -15,6 +15,7 @@ end
 RSpec.configure do |config|
   config.include(Warden::Spec::Helpers)
   config.include(Warden::Test::Helpers)
+  config.include(Warden::Test::Mock)
 
   def load_strategies
     Dir[File.join(File.dirname(__FILE__), "helpers", "strategies", "**/*.rb")].each do |f|
